backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
redis
httpx
pydantic-settings
psycopg2-binary
python-multipart
